--- research/second_llm_pass/ds_gen/example_config.json
@@ -1,20 +1,0 @@
-{
-  "frames": [
-    {
-      "content": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/lidl/Kopia test_data_2024_11_25_lidl_plus_content_generic_2024-12-05T07_39_49.726955559+01_00.csv",
-      "coupons": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/lidl/Kopia test_data_2024_11_25_lidl_plus_cupons_2024-12-05T07_54_54.63740938+01_00.csv",
-      "format": 1
-    },
-    {
-      "content": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/rossmann/Kopia test_data_2024_03_07_rossmann_content_generic_2024-12-05T10_24_07.981399375+01_00.csv",
-      "coupons": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/rossmann/Kopia test_data_2024_03_07_rossmann_coupons_2024-12-05T10_24_55.372924301+01_00.csv",
-      "format": 1
-    },
-    {
-      "content": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/dm/Kopia test_data_2024_03_07_dm_content_generic_2024-12-05T10_09_32.502568365+01_00.csv",
-      "coupons": "/home/gustaw/PycharmProjects/ZPP_Murmuras/datasets/coupons_1-20250224T131740Z-001/coupons_1/dm/Kopia test_data_2024_03_07_dm_coupons_2024-12-05T10_11_37.906933437+01_00.csv",
-      "format": 2
-    }
-  ],
-  "json_format": true
-}
